soc: add a new version of LunaSoC without litex-bios and vendored lambdasoc peripherals
