+++ docs/requirements.txt
@@ -7,5 +7,4 @@ jinja2==3.1.3
 amaranth-stdio @ git+https://github.com/amaranth-lang/amaranth-stdio@4a14bb17
 amaranth-soc @ git+https://github.com/amaranth-lang/amaranth-soc@d2185e3
 minerva @ git+https://github.com/minerva-cpu/minerva
-lambdasoc @ git+https://github.com/antoinevg/lambdasoc.git@antoinevg/support-vexriscv
 luna @ git+https://github.com/greatscottgadgets/luna@main
