+++ luna_soc/gateware/wishbone/__init__.py
@@ -4,8 +4,6 @@
 # Copyright (c) 2023 Great Scott Gadgets <info@greatscottgadgets.com>
 # SPDX-License-Identifier: BSD-3-Clause
 
-from .base import *
-from .gpio import *
-from .led  import *
-from .uart import *
-from .usb2 import *
+from .memory import *
+from .qspi   import *
+from .sram   import *
