+++ luna_soc/generate/gensvd.py
@@ -6,24 +6,23 @@
 
 """Generate a SVD file for SoC designs."""
 
-import amaranth
-
-import amaranth_soc
-from  amaranth_soc.memory import MemoryMap, ResourceInfo
-
-from lambdasoc.soc.cpu import CPUSoC
+from os import path
+import logging
+import sys
 
 from xml.dom import minidom
 from xml.etree import ElementTree
 from xml.etree.ElementTree import Element, SubElement, Comment, tostring
 
-from os import path
-import logging
-import sys
+from  amaranth_soc.memory import MemoryMap, ResourceInfo
+import amaranth_soc
+
+from ..gateware.lunasoc import LunaSoC
+
 
 class GenSVD:
 
-    def __init__(self, soc: CPUSoC):
+    def __init__(self, soc: LunaSoC):
         self._soc = soc
 
 
@@ -37,9 +36,13 @@ class GenSVD:
         # <peripherals />
         peripherals = SubElement(device, "peripherals")
 
+        # TODO this should be determined by introspection
+        memories = ["bootrom", "scratchpad", "mainram", "ram", "rom", "spiflash"]
+
         window: MemoryMap
         for window, (start, stop, ratio) in self._soc.memory_map.windows():
-            if window.name in ["bootrom", "scratchpad", "mainram", "ram", "rom"]:
+            print(window.name)
+            if window.name in memories:
                 logging.debug("Skipping non-peripheral resource: {}".format(window.name))
                 continue
 
@@ -59,7 +62,7 @@ class GenSVD:
 
         window: MemoryMap
         for window, (start, stop, ratio) in self._soc.memory_map.windows():
-            if window.name not in ["bootrom", "scratchpad", "mainram", "ram", "rom"]:
+            if window.name not in memories:
                 continue
 
             memoryRegion = SubElement(memoryRegions, "memoryRegion")
@@ -86,7 +89,7 @@ class GenSVD:
 
 # - section helpers -----------------------------------------------------------
 
-def _generate_section_device(soc: CPUSoC, vendor, name, description):
+def _generate_section_device(soc: LunaSoC, vendor, name, description):
     device = Element("device")
     device.set("schemaVersion", "1.1")
     device.set("xmlns:xs", "http://www.w3.org/2001/XMLSchema-instance")
@@ -146,6 +149,8 @@ def _generate_section_peripheral(peripherals: Element, soc, window: MemoryMap, s
 
 def _generate_section_register(registers: Element, window: MemoryMap, resource_info: ResourceInfo):
     resource: amaranth_soc.csr.bus.Element = resource_info.resource
+    print(f"resource={type(resource)}  resource_info={type(resource_info)} ")
+    print(f"start={resource_info.start} end={resource_info.end}")
     assert type(resource) == amaranth_soc.csr.bus.Element
     from amaranth_soc.csr.bus import Element
